Another try

+++ postbuild.js
@@ -2,13 +2,17 @@ const { glob } = require('glob');
 const { rimraf } = require('rimraf');
 
 const start = async () => {
-  await glob('release/*').then((files) => {
-    files.forEach((file) => {
-      if ((!file.includes('Tax-Helper-Setup') && !file.endsWith('exe')) || file.endsWith('blockmap')) {
-        rimraf.sync(file);
-      }
+  const globPromise = glob('release/*');
+  console.log('🚀 ~ file: postbuild.js:6 ~ start ~ globPromise:', globPromise);
+  if (globPromise) {
+    globPromise.then((files) => {
+      files.forEach((file) => {
+        if ((!file.includes('Tax-Helper-Setup') && !file.endsWith('exe')) || file.endsWith('blockmap')) {
+          rimraf.sync(file);
+        }
+      });
     });
-  });
+  }
 };
 
 start();
